--- tests/integration/helpers.py
@@ -1,386 +1,0 @@
-# Copyright 2022 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-import itertools
-import secrets
-import string
-from typing import Dict, List, Optional
-
-import yaml
-from juju.model import Model
-from juju.unit import Unit
-from mysql.connector.errors import (
-    DatabaseError,
-    InterfaceError,
-    OperationalError,
-    ProgrammingError,
-)
-from pytest_operator.plugin import OpsTest
-from tenacity import retry, stop_after_attempt, wait_fixed
-
-from constants import CONTAINER_NAME, SERVER_CONFIG_USERNAME
-
-from . import juju_
-from .connector import MySQLConnector
-
-
-def generate_random_string(length: int) -> str:
-    """Generate a random string of the provided length.
-
-    Args:
-        length: the length of the random string to generate
-
-    Returns:
-        A random string comprised of letters and digits
-    """
-    choices = string.ascii_letters + string.digits
-    return "".join([secrets.choice(choices) for i in range(length)])
-
-
-async def get_unit_address(
-    ops_test: OpsTest, unit_name: str, model: Optional[Model] = None
-) -> str:
-    """Get unit IP address.
-
-    Args:
-        ops_test: The ops test framework instance
-        unit_name: The name of the unit
-        model: (Optional) model to use instead of ops_test.model
-
-    Returns:
-        IP address of the unit
-    """
-    if model is None:
-        model = ops_test.model
-    status = await model.get_status()
-    return status["applications"][unit_name.split("/")[0]].units[unit_name]["address"]
-
-
-async def get_cluster_status(unit: Unit, cluster_set=False) -> Dict:
-    """Get the cluster status by running the get-cluster-status action.
-
-    Args:
-        unit: The unit on which to execute the action on
-        cluster_set: A boolean indicating whether to get the cluster set status
-
-    Returns:
-        A dictionary representing the cluster status
-    """
-    if cluster_set:
-        results = await juju_.run_action(
-            unit, "get-cluster-status", **{"--wait": "5m", "cluster-set": True}
-        )
-    else:
-        results = await juju_.run_action(unit, "get-cluster-status")
-    return results.get("status", {})
-
-
-async def get_relation_data(
-    ops_test: OpsTest,
-    application_name: str,
-    relation_name: str,
-) -> list:
-    """Returns a list that contains the relation-data.
-
-    Args:
-        ops_test: The ops test framework instance
-        application_name: The name of the application
-        relation_name: name of the relation to get connection data from
-    Returns:
-        a list that contains the relation-data
-    """
-    # get available unit id for the desired application
-    unit_names = [
-        app_unit.name for app_unit in ops_test.model.applications[application_name].units
-    ]
-    assert len(unit_names) > 0
-    unit_name = unit_names[0]
-
-    raw_data = (await ops_test.juju("show-unit", unit_name))[1]
-    assert raw_data, f"no unit info could be grabbed for {unit_name}"
-
-    data = yaml.safe_load(raw_data)
-    # Filter the data based on the relation name.
-    relation_data = [v for v in data[unit_name]["relation-info"] if v["endpoint"] == relation_name]
-    assert relation_data, (
-        f"no relation data could be grabbed on relation with endpoint {relation_name}"
-    )
-
-    return relation_data
-
-
-async def get_primary_unit(
-    ops_test: OpsTest,
-    unit: Unit,
-    app_name: str,
-) -> Optional[Unit]:
-    """Helper to retrieve the primary unit.
-
-    Args:
-        ops_test: The ops test object passed into every test case
-        unit: A unit on which to execute commands/queries/actions on
-        app_name: The name of the test application
-
-    Returns:
-        A juju unit that is a MySQL primary
-    """
-    cluster_status = await get_cluster_status(unit)
-
-    primary_label = next(
-        iter([
-            label
-            for label, member in cluster_status["defaultreplicaset"]["topology"].items()
-            if member["mode"] == "r/w"
-        ])
-    )
-    primary_name = "/".join(primary_label.rsplit("-", 1))
-
-    for unit in ops_test.model.applications[app_name].units:
-        if unit.name == primary_name:
-            return unit
-
-    return None
-
-
-def execute_queries_on_unit(
-    unit_address: str,
-    username: str,
-    password: str,
-    queries: List[str],
-    commit: bool = False,
-    raw: bool = False,
-) -> List:
-    """Execute given MySQL queries on a unit.
-
-    Args:
-        unit_address: The public IP address of the unit to execute the queries on
-        username: The MySQL username
-        password: The MySQL password
-        queries: A list of queries to execute
-        commit: A keyword arg indicating whether there are any writes queries
-        raw: Whether MySQL results are returned as is, rather than converted to Python types.
-
-    Returns:
-        A list of rows that were potentially queried
-    """
-    config = {
-        "user": username,
-        "password": password,
-        "host": unit_address,
-        "raise_on_warnings": False,
-        "raw": raw,
-    }
-
-    with MySQLConnector(config, commit) as cursor:
-        for query in queries:
-            cursor.execute(query)
-        output = list(itertools.chain(*cursor.fetchall()))
-
-    return output
-
-
-async def get_server_config_credentials(unit: Unit) -> Dict:
-    """Helper to run an action to retrieve server config credentials.
-
-    Args:
-        unit: The juju unit on which to run the get-password action for server-config credentials
-
-    Returns:
-        A dictionary with the server config username and password
-    """
-    return await juju_.run_action(unit, "get-password", username=SERVER_CONFIG_USERNAME)
-
-
-async def fetch_credentials(unit: Unit, username: str = "") -> Dict:
-    """Helper to run an action to fetch credentials.
-
-    Args:
-        unit: The juju unit on which to run the get-password action for credentials
-        username: The username for which to fetch credentials
-
-    Returns:
-        A dictionary with the server config username and password
-    """
-    if not username:
-        return await juju_.run_action(unit, "get-password")
-    else:
-        return await juju_.run_action(unit, "get-password", username=username)
-
-
-async def rotate_credentials(unit: Unit, username: str = "", password: str = "") -> Dict:
-    """Helper to run an action to rotate credentials.
-
-    Args:
-        unit: The juju unit on which to run the set-password action for credentials
-        username: The username for which to rotate credentials
-        password: The password with which to rotate credentials
-
-    Returns:
-        A dictionary with the action result
-    """
-    if not username:
-        return await juju_.run_action(unit, "set-password")
-    elif not password:
-        return await juju_.run_action(unit, "set-password", username=username)
-    else:
-        return await juju_.run_action(unit, "set-password", username=username, password=password)
-
-
-async def scale_application(
-    ops_test: OpsTest, application_name: str, desired_count: int, wait: bool = True
-) -> None:
-    """Scale a given application to the desired unit count.
-
-    Args:
-        ops_test: The ops test framework
-        application_name: The name of the application
-        desired_count: The number of units to scale to
-        wait: Boolean indicating whether to wait until units
-            reach desired count
-    """
-    await ops_test.model.applications[application_name].scale(desired_count)
-
-    if desired_count > 0 and wait:
-        async with ops_test.fast_forward("60s"):
-            await ops_test.model.wait_for_idle(
-                apps=[application_name],
-                status="active",
-                timeout=(15 * 60),
-                wait_for_exact_units=desired_count,
-                raise_on_blocked=True,
-                raise_on_error=False,
-            )
-
-
-def is_relation_joined(
-    ops_test: OpsTest,
-    endpoint_one: str,
-    endpoint_two: str,
-    application_one: Optional[str] = None,
-    application_two: Optional[str] = None,
-) -> bool:
-    """Check if a relation is joined.
-
-    Args:
-        ops_test: The ops test object passed into every test case
-        endpoint_one: The first endpoint of the relation
-        endpoint_two: The second endpoint of the relation
-        application_one: The name of the first application
-        application_two: The name of the second application
-    """
-    for rel in ops_test.model.relations:
-        if application_one and application_two:
-            endpoints = [
-                f"{endpoint.application_name}:{endpoint.name}" for endpoint in rel.endpoints
-            ]
-            if (
-                f"{application_one}:{endpoint_one}" in endpoints
-                and f"{application_two}:{endpoint_two}" in endpoints
-            ):
-                return True
-        else:
-            endpoints = [endpoint.name for endpoint in rel.endpoints]
-            if endpoint_one in endpoints and endpoint_two in endpoints:
-                return True
-    return False
-
-
-def is_relation_broken(ops_test: OpsTest, endpoint_one: str, endpoint_two: str) -> bool:
-    """Check if a relation is broken.
-
-    Args:
-        ops_test: The ops test object passed into every test case
-        endpoint_one: The first endpoint of the relation
-        endpoint_two: The second endpoint of the relation
-    """
-    for rel in ops_test.model.relations:
-        endpoints = [endpoint.name for endpoint in rel.endpoints]
-        if endpoint_one not in endpoints and endpoint_two not in endpoints:
-            return True
-    return False
-
-
-async def app_name(ops_test: OpsTest) -> str:
-    """Returns the name of the application running MySQL.
-
-    This is important since not all deployments of the MySQL charm have the application name
-    "mysql-k8s".
-
-    Note: if multiple clusters are running MySQL this will return the one first found.
-    """
-    status = await ops_test.model.get_status()
-    for app in ops_test.model.applications:
-        # note that format of the charm field is not exactly "mysql-k8s" but instead takes the form
-        # of `local:focal/mysql-6`
-        if "mysql-k8s" in status["applications"][app]["charm"]:
-            return app
-
-    return None
-
-
-@retry(stop=stop_after_attempt(8), wait=wait_fixed(15), reraise=True)
-def is_connection_possible(credentials: Dict, **extra_opts) -> bool:
-    """Test a connection to a MySQL server.
-
-    Args:
-        credentials: A dictionary with the credentials to test
-        extra_opts: extra options for mysql connection
-    """
-    config = {
-        "user": credentials["username"],
-        "password": credentials["password"],
-        "host": credentials["host"],
-        "raise_on_warnings": False,
-        "connection_timeout": 10,
-        **extra_opts,
-    }
-
-    try:
-        with MySQLConnector(config) as cursor:
-            cursor.execute("SELECT 1")
-            return cursor.fetchone()[0] == 1
-    except (DatabaseError, InterfaceError, OperationalError, ProgrammingError):
-        # Errors raised when the connection is not possible
-        return False
-
-
-async def retrieve_database_variable_value(
-    ops_test: OpsTest, unit: Unit, variable_name: str
-) -> str:
-    """Retrieve a database variable value as a string.
-
-    Args:
-        ops_test: The ops test framework instance
-        unit: The unit to retrieve the variable
-        variable_name: The name of the variable to retrieve
-    Returns:
-        The variable value (str)
-    """
-    unit_ip = await get_unit_address(ops_test, unit.name)
-
-    server_config_creds = await get_server_config_credentials(unit)
-    queries = [f"SELECT @@{variable_name};"]
-
-    output = execute_queries_on_unit(
-        unit_ip, server_config_creds["username"], server_config_creds["password"], queries
-    )
-
-    return output[0]
-
-
-async def start_mysqld_exporter(ops_test: OpsTest, unit: Unit) -> None:
-    """Start mysqld exporter pebble service on the provided unit.
-
-    Args:
-        ops_test: The ops test framework
-        unit: The unit to start mysqld exporter on
-    """
-    await ops_test.juju(
-        "ssh",
-        "--container",
-        CONTAINER_NAME,
-        unit.name,
-        "pebble",
-        "start",
-        "mysqld_exporter",
-    )
