+++ tests/integration/helpers_ha.py
@@ -2,7 +2,10 @@
 # Copyright 2025 Canonical Ltd.
 # See LICENSE file for licensing details.
 
+import itertools
 import json
+import secrets
+import string
 import subprocess
 from collections.abc import Callable, Generator
 from contextlib import contextmanager
@@ -16,6 +19,12 @@ from jubilant_backports.statustypes import Status
 from lightkube.core.client import Client
 from lightkube.resources.apps_v1 import StatefulSet
 from lightkube.resources.core_v1 import Endpoints, PersistentVolume, PersistentVolumeClaim, Pod
+from mysql.connector.errors import (
+    DatabaseError,
+    InterfaceError,
+    OperationalError,
+    ProgrammingError,
+)
 from tenacity import (
     Retrying,
     retry,
@@ -30,7 +39,7 @@ from constants import (
     SERVER_CONFIG_USERNAME,
 )
 
-from .helpers import execute_queries_on_unit
+from .connector import MySQLConnector
 
 CHARM_METADATA = yaml.safe_load(Path("./metadata.yaml").read_text())
 
@@ -39,6 +48,19 @@ TEST_DATABASE_NAME = "testing"
 
 JujuModelStatusFn = Callable[[Status], bool]
 JujuAppsStatusFn = Callable[[Status, str], bool]
+
+
+def generate_random_string(length: int) -> str:
+    """Generate a random string of the provided length.
+
+    Args:
+        length: the length of the random string to generate
+
+    Returns:
+        A random string comprised of letters and digits
+    """
+    choices = string.ascii_letters + string.digits
+    return "".join([secrets.choice(choices) for i in range(length)])
 
 
 def check_mysql_instances_online(
@@ -672,3 +694,66 @@ def wait_for_unit_message(app_name: str, unit_name: str, unit_message: str) -> J
     return lambda status: (
         status.apps[app_name].units[unit_name].workload_status.message == unit_message
     )
+
+
+def execute_queries_on_unit(
+    unit_address: str,
+    username: str,
+    password: str,
+    queries: list[str],
+    commit: bool = False,
+    raw: bool = False,
+) -> list:
+    """Execute given MySQL queries on a unit.
+
+    Args:
+        unit_address: The public IP address of the unit to execute the queries on
+        username: The MySQL username
+        password: The MySQL password
+        queries: A list of queries to execute
+        commit: A keyword arg indicating whether there are any writes queries
+        raw: Whether MySQL results are returned as is, rather than converted to Python types.
+
+    Returns:
+        A list of rows that were potentially queried
+    """
+    config = {
+        "user": username,
+        "password": password,
+        "host": unit_address,
+        "raise_on_warnings": False,
+        "raw": raw,
+    }
+
+    with MySQLConnector(config, commit) as cursor:
+        for query in queries:
+            cursor.execute(query)
+        output = list(itertools.chain(*cursor.fetchall()))
+
+    return output
+
+
+@retry(stop=stop_after_attempt(8), wait=wait_fixed(15), reraise=True)
+def is_connection_possible(credentials: dict, **extra_opts) -> bool:
+    """Test a connection to a MySQL server.
+
+    Args:
+        credentials: A dictionary with the credentials to test
+        extra_opts: extra options for mysql connection
+    """
+    config = {
+        "user": credentials["username"],
+        "password": credentials["password"],
+        "host": credentials["host"],
+        "raise_on_warnings": False,
+        "connection_timeout": 10,
+        **extra_opts,
+    }
+
+    try:
+        with MySQLConnector(config) as cursor:
+            cursor.execute("SELECT 1")
+            return cursor.fetchone()[0] == 1
+    except (DatabaseError, InterfaceError, OperationalError, ProgrammingError):
+        # Errors raised when the connection is not possible
+        return False
