+++ tests/integration/integration/high_availability/test_replication_data_consistency.py
@@ -7,10 +7,10 @@ import jubilant_backports
 import pytest
 from jubilant_backports import Juju
 
-from ...helpers import generate_random_string
 from ...helpers_ha import (
     CHARM_METADATA,
     check_mysql_units_writes_increment,
+    generate_random_string,
     insert_mysql_test_data,
     remove_mysql_test_data,
     verify_mysql_test_data,
