+++ tests/integration/integration/high_availability/test_self_healing_pod.py
@@ -7,10 +7,10 @@ import jubilant_backports
 import pytest
 from jubilant_backports import Juju
 
-from ...helpers import generate_random_string
 from ...helpers_ha import (
     CHARM_METADATA,
     delete_k8s_pod,
+    generate_random_string,
     get_app_units,
     insert_mysql_test_data,
     remove_mysql_test_data,
